requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
selenium==4.16.0
webdriver-manager==4.0.1